Review feedback

+++ embrace-android-delivery/src/main/kotlin/io/embrace/android/embracesdk/internal/delivery/scheduling/SchedulingServiceImpl.kt
@@ -5,8 +5,9 @@ package io.embrace.android.embracesdk.internal.delivery.scheduling
 import io.embrace.android.embracesdk.internal.clock.Clock
 import io.embrace.android.embracesdk.internal.comms.api.ApiResponse
 import io.embrace.android.embracesdk.internal.comms.api.Endpoint
+import io.embrace.android.embracesdk.internal.delivery.StoredTelemetryMetadata
 import io.embrace.android.embracesdk.internal.delivery.execution.RequestExecutionService
-import io.embrace.android.embracesdk.internal.storage.PayloadReference
+import io.embrace.android.embracesdk.internal.delivery.storedTelemetryComparator
 import io.embrace.android.embracesdk.internal.storage.StorageService2
 import io.embrace.android.embracesdk.internal.worker.BackgroundWorker
 import java.io.InputStream
@@ -27,34 +28,47 @@ internal class SchedulingServiceImpl(
 
     private val blockedEndpoints: MutableMap<Endpoint, Long> = ConcurrentHashMap()
     private val sendLoopActive = AtomicBoolean(false)
-    private val activeSends: MutableSet<PayloadReference> = Collections.synchronizedSet(HashSet())
-    private val payloadsToRetry: MutableMap<PayloadReference, RetryInstance> = ConcurrentHashMap()
+    private val queryForPayloads = AtomicBoolean(true)
+    private val activeSends: MutableSet<StoredTelemetryMetadata> = Collections.newSetFromMap(ConcurrentHashMap())
+    private val payloadsToRetry: MutableMap<StoredTelemetryMetadata, RetryInstance> = ConcurrentHashMap()
 
     override fun onPayloadIntake() {
-        // When a payload arrives, check to see if there's already an active job try to deliver payloads
-        // If not, schedule job. If so, do nothing.
-        if (sendLoopActive.compareAndSet(false, true)) {
-            schedulingWorker.submit {
-                queueLoop()
-            }
-        }
+        queryForPayloads.set(true)
+        startDeliveryLoop()
     }
 
     override fun handleCrash(crashId: String) {
         // TODO: get ready to die
     }
 
+    private fun startDeliveryLoop() {
+        // When a payload arrives, check to see if there's already an active job try to deliver payloads
+        // If not, schedule job. If so, do nothing.
+        if (sendLoopActive.compareAndSet(false, true)) {
+            schedulingWorker.submit {
+                deliveryLoop()
+            }
+        }
+    }
+
     /**
      * Loop through the payloads ready to be sent by priority and queue for delivery
      */
-    private fun queueLoop() {
+    private fun deliveryLoop() {
         try {
-            var readyPayloads = getReadyPayloads()
-            while (readyPayloads.isNotEmpty() && readyToSend()) {
-                readyPayloads.forEach { payload ->
-                    queueSend(payload)
+            var deliveryQueue = getDeliveryQueue()
+            while (deliveryQueue.isNotEmpty() && readyToSend()) {
+                deliveryQueue.poll()?.let { payload ->
+                    queueDelivery(payload)
                 }
-                readyPayloads = getReadyPayloads()
+
+                if (queryForPayloads.compareAndSet(true, false)) {
+                    deliveryQueue = getDeliveryQueue()
+                }
+
+                if (deliveryQueue.isEmpty()) {
+                    deliveryQueue = getDeliveryQueue()
+                }
             }
         } finally {
             sendLoopActive.set(false)
@@ -62,25 +76,21 @@ internal class SchedulingServiceImpl(
         }
     }
 
-    private fun getReadyPayloads(): List<PayloadReference> {
-        val payloads = storageService.getPayloadsByPriority()
-        val readyPayloads = LinkedList<PayloadReference>()
-        payloads.forEach { payload ->
-            if (payload.shouldSendPayload()) {
-                readyPayloads.push(payload)
-            }
-        }
-        return readyPayloads
-    }
+    private fun getDeliveryQueue() = LinkedList(getReadyPayloads())
 
-    private fun PayloadReference.shouldSendPayload(): Boolean {
+    private fun getReadyPayloads(): List<StoredTelemetryMetadata> =
+        storageService.getPayloadsByPriority()
+            .filter { it.shouldSendPayload() }
+            .sortedWith(storedTelemetryComparator)
+
+    private fun StoredTelemetryMetadata.shouldSendPayload(): Boolean {
         // determine if the given payload is eligible to be sent
         // i.e. not already being sent, endpoint not blocked by 429, and isn't waiting to be retried
-        updateBlockedEndpoint(endpoint)
+        updateBlockedEndpoint(envelopeType.endpoint)
 
         return if (activeSends.contains(this)) {
             false
-        } else if (blockedEndpoints.containsKey(endpoint)) {
+        } else if (blockedEndpoints.containsKey(envelopeType.endpoint)) {
             false
         } else {
             payloadsToRetry[this]?.run {
@@ -89,30 +99,43 @@ internal class SchedulingServiceImpl(
         }
     }
 
-    private fun queueSend(payload: PayloadReference): Future<ApiResponse> {
+    private fun queueDelivery(payload: StoredTelemetryMetadata): Future<ApiResponse> {
         activeSends.add(payload)
         return deliveryWorker.submit<ApiResponse> {
-            executionService.attemptHttpRequest(
-                payloadStream = payload.toInputStream(),
-                envelopeType = payload.envelopeType
-            ).apply {
-                if (!shouldRetry) {
-                    payloadsToRetry.remove(payload)
-                    storageService.deletePayload(payload)
-                } else {
-                    val failedAttempts = payloadsToRetry[payload]?.let { it.failedAttempts + 1 } ?: 1
-                    payloadsToRetry[payload] = RetryInstance(
-                        failedAttempts = failedAttempts,
-                        nextRetryTimeMs = calculateNextRetryTime(retryCount = failedAttempts - 1)
-                    )
-                }
+            try {
+                val payloadStream = payload.toStream()
+                if (payloadStream != null) {
+                    executionService.attemptHttpRequest(
+                        payloadStream = { payloadStream },
+                        envelopeType = payload.envelopeType
+                    ).apply {
+                        if (!shouldRetry) {
+                            // If the response is such that we should not ever retry the delivery of this payload,
+                            // delete it from both the in memory retry payloads map and on disk
+                            payloadsToRetry.remove(payload)
+                            storageService.deletePayload(payload)
+                        } else {
+                            // If delivery of this payload should be retried, add or replace the entry in the retry map
+                            // with the new values for how many times it has failed, and when the next retry should happen
+                            val retryAttempts = payloadsToRetry[payload]?.failedAttempts ?: 0
+                            val nextRetryTimeMs = calculateNextRetryTime(retryAttempts = retryAttempts)
+                            payloadsToRetry[payload] = RetryInstance(
+                                failedAttempts = retryAttempts + 1,
+                                nextRetryTimeMs = nextRetryTimeMs
+                            )
+                        }
 
-                if (this is ApiResponse.TooManyRequests) {
-                    retryAfter?.let { delayMs ->
-                        blockedEndpoints[endpoint] = clock.now() + delayMs
+                        if (this is ApiResponse.TooManyRequests) {
+                            retryAfter?.let { delayMs ->
+                                blockedEndpoints[endpoint] = clock.now() + delayMs
+                            }
+                        }
                     }
+                } else {
+                    // Could not find payload. Do not retry.
+                    ApiResponse.NoPayload
                 }
-
+            } finally {
                 activeSends.remove(payload)
             }
         }
@@ -121,10 +144,10 @@ internal class SchedulingServiceImpl(
     private fun scheduleNextCheck() {
         payloadsToRetry.map { it.value.nextRetryTimeMs }.minOrNull()?.let { timestampMs ->
             if (timestampMs <= clock.now()) {
-                onPayloadIntake()
+                startDeliveryLoop()
             } else if (timestampMs != Long.MAX_VALUE) {
                 schedulingWorker.schedule<Unit>(
-                    ::onPayloadIntake,
+                    ::startDeliveryLoop,
                     calculateDelay(timestampMs),
                     TimeUnit.MILLISECONDS
                 )
@@ -145,11 +168,11 @@ internal class SchedulingServiceImpl(
         }
     }
 
-    private fun PayloadReference.toInputStream(): () -> InputStream? = { storageService.loadPayloadAsStream(this) }
+    private fun StoredTelemetryMetadata.toStream(): InputStream? = storageService.loadPayloadAsStream(this)
 
     private fun calculateDelay(nextRetryTimeMs: Long): Long = nextRetryTimeMs - clock.now()
 
-    private fun calculateNextRetryTime(retryCount: Int): Long = clock.now() + (INITIAL_DELAY_MS * (1 shl retryCount))
+    private fun calculateNextRetryTime(retryAttempts: Int): Long = clock.now() + (INITIAL_DELAY_MS * (1 shl retryAttempts))
 
     private data class RetryInstance(
         val failedAttempts: Int,
